+++ src/5 Elastic Solids/p52.jl
@@ -15,8 +15,8 @@ function p52(data::Dict)
   
   # Handle :r direction (implicit axisymmetric)
   if struc_el.direction == :r
-    nre = struc_el.nxe::Int64
-    nze = struc_el.nye::Int64
+    nre = struc_el.nxe::Int
+    nze = struc_el.nye::Int
   end
   
   # Add radial stress
@@ -62,7 +62,7 @@ function p52(data::Dict)
     exit(1)
   end
     
-  nf = ones(Int64, nodof, nn)
+  nf = ones(Int, nodof, nn)
   if :support in keys(data)
     for i in 1:size(data[:support], 1)
       nf[:, data[:support][i][1]] = data[:support][i][2]
@@ -92,7 +92,7 @@ function p52(data::Dict)
     r_coords = data[:r_coords]
   end
 
-  etype = ones(Int64, nels)
+  etype = ones(Int, nels)
   if :etype in keys(data)
     etype = data[:etype]
   end
@@ -100,8 +100,8 @@ function p52(data::Dict)
   @assert :lth in keys(data)
   @assert :iflag in keys(data)
   @assert :chi in keys(data)
-  lth = data[:lth]::Int64
-  iflag = data[:iflag]::Int64
+  lth = data[:lth]::Int
+  iflag = data[:iflag]::Int
   chi = data[:chi]*pi/180.0
   ca = cos(chi)
   sa = sin(chi)
@@ -110,13 +110,13 @@ function p52(data::Dict)
   # All other arrays
   
   points = zeros(struc_el.nip, ndim)
-  g = zeros(Int64, ndof)
+  g = zeros(Int, ndof)
   g_coord = zeros(ndim,nn)
   fun = zeros(fin_el.nod)
   coord = zeros(fin_el.nod, ndim)
   gamma = zeros(nels)
   jac = zeros(ndim, ndim)
-  g_num = zeros(Int64, fin_el.nod, nels)
+  g_num = zeros(Int, fin_el.nod, nels)
   der = zeros(ndim, fin_el.nod)
   deriv = zeros(ndim, fin_el.nod)
   bee = zeros(nst,ndof)
@@ -126,8 +126,8 @@ function p52(data::Dict)
   kg = zeros(ndof, ndof)
   eld = zeros(ndof)
   weights = zeros(struc_el.nip)
-  g_g = zeros(Int64, ndof, nels)
-  num = zeros(Int64, fin_el.nod)
+  g_g = zeros(Int, ndof, nels)
+  num = zeros(Int, fin_el.nod)
   actions = zeros(ndof, nels)
   displacements = zeros(size(nf, 1), ndim)
   gc = ones(ndim)
@@ -137,7 +137,7 @@ function p52(data::Dict)
   
   formnf!(nodof, nn, nf)
   neq = maximum(nf)
-  kdiag = zeros(Int64, neq)
+  kdiag = zeros(Int, neq)
   
   # Find global array sizes
   
@@ -192,9 +192,9 @@ function p52(data::Dict)
   if :fixed_freedoms in keys(data)
     fixed_freedoms = size(data[:fixed_freedoms], 1)
   end
-  no = zeros(Int64, fixed_freedoms)
-  node = zeros(Int64, fixed_freedoms)
-  sense = zeros(Int64, fixed_freedoms)
+  no = zeros(Int, fixed_freedoms)
+  node = zeros(Int, fixed_freedoms)
+  sense = zeros(Int, fixed_freedoms)
   value = zeros(Float64, fixed_freedoms)
   if :fixed_freedoms in keys(data) && fixed_freedoms > 0
     for i in 1:fixed_freedoms
